change size_t to int

+++ include/core/vertex_set.cuh
@@ -16,7 +16,7 @@ concept IsVertexSetImpl = requires(Impl t, VIndex_t *data, VIndex_t size) {
     { t.init(data, size) } -> std::same_as<void>;
     { t.size() } -> std::same_as<VIndex_t>;
     { t.data() } -> std::same_as<VIndex_t *>;
-    { t.storage_space() } -> std::same_as<size_t>;
+    { t.storage_space() } -> std::same_as<int>;
     { t.clear() } -> std::same_as<void>;
     { t.intersect(t, t) } -> std::same_as<void>;
     // TODO: 增加 for each
